First commit
First commit, which include everything developped so far. Most of the work as gone on setting up a simple console exemple, to test things out.
Work still needs to be done on the abstraction (mostly for the units, such as the Mecha class) and generalization (allowing for several ressources to be harvested, rotating the conveyors, selecting multiple blueprint for the builders, ...).
But all in all, it's a functionnal demo of the very basic concept for the game, and that's great !