Cleaning up the refactoring and correcting a few bugs
The build is back to it's state prior to the refactoring, with the inclusion of the grabber.
Work still needs to be done to ensure everything is properly decoupled and to remove a number of arguments that get passed around when it's not necessary (the parent of the FactoryComponent in the constructor comes to mind).

It will be necessary to work on the Ressource-handling part of the components; right now everything points back to a Ressource object, despite most of the classes needing only a RessourceTemplate and the actual Ressource being handled by the Container component.

A few events will have to be added to the components and a proper way to automaticly adding/removing references still has to be worked out.

A way to know the percent of advancement for a given task (generator & producer) would be nice

Oh, and by the way; we really need to be consistent in the terms we use. Is it generator or harvester ? Builder, Constructor, Producer, ... ?